pyproject.toml
[build-system]
requires = ["setuptools>=64.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "memorypack"
version = "0.1.0"
description = "Compress markdown knowledge bases into context-efficient format for LLMs"
requires-python = ">=3.9"
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "python-frontmatter>=1.0",
    "sentence-transformers>=2.2",
    "transformers>=4.30",
    "torch>=2.0",
    "scikit-learn>=1.3",
    "nltk>=3.8",
    "numpy>=1.24",
]

[project.scripts]
memorypack = "memorypack.cli:main"

[tool.setuptools.packages.find]
include = ["memorypack*"]
